Remove useShowBlockTools and combine block-tools selectors
